+++ src/main/java/org/jenkinsci/plugins/DependencyTrack/model/Thresholds.java
@@ -33,10 +33,12 @@ public class Thresholds implements Serializable {
         public Integer unstableHigh;
         public Integer unstableMedium;
         public Integer unstableLow;
+        public Integer unstableUnassigned;
         public Integer failedCritical;
         public Integer failedHigh;
         public Integer failedMedium;
         public Integer failedLow;
+        public Integer failedUnassigned;
     }
 
     @EqualsAndHashCode
@@ -46,9 +48,11 @@ public class Thresholds implements Serializable {
         public Integer unstableHigh;
         public Integer unstableMedium;
         public Integer unstableLow;
+        public Integer unstableUnassigned;
         public Integer failedCritical;
         public Integer failedHigh;
         public Integer failedMedium;
         public Integer failedLow;
+        public Integer failedUnassigned;
     }
 }